backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
beautifulsoup4
requests
langchain
langchain-community
langchain-core
langchain-google-genai>=1.0.3
google-generativeai>=0.5.2

python-dotenv
pydantic
